Set up a skeletal Terraform configuration for production infra
State backend and AWS provider.  A single configuration for now but a
layout amenable to the expectation of multiple configurations in the
future.

The directory name is "env", not "terraform", because I expect these
directories to house additional non-Terraform files needed for each
environment.  Snippets shared between the configurations can then live
directly under env/ and be symlinked.

Multiple configurations using local modules are the primary and
recommended¹ way to have multiple environments that are similar but not
identical.  Terraform "workspaces" are another way manage multiple
environments, but they're intended for multiple instances of a single
configuration.²  You can parameterize the configuration on the workspace
name, but it's harder to express larger scale divergence (different
resources, different modules) with just parameterization due to the lack
of any top-level conditional constructs.

Cribbed-from: <https://github.com/nextstrain/nextstrain.org/commit/5921be61>

¹ c.f. <https://developer.hashicorp.com/terraform/tutorials/modules/organize-configuration>
² <https://developer.hashicorp.com/terraform/cli/workspaces#use-cases>