Inline the permissions policy for GitHubActionsRoleNextstrainTmpBucket
No need to have an extra managed policy floating around when it's not
used by any other principal.

I'll remove the managed policy in the subsequent commit.  Terraform
seems to require a two-step apply here to get the right ordering of
detaching the managed policy from the role and then destroying the
managed policy.