Special-case forecasts-ncov to give it read-only access to nextstrain-ncov-private
It needs read-only access to (at least)
<s3://nextstrain-ncov-private/metadata.tsv.zst>, and special-casing it
is the quick and easy way to do that.

The forecasts-ncov repo is conceptually-speaking maybe best thought of
as a forecasts/ directory within the ncov repo, similar to how the
ncov-ingest repo is best thought of as an ingest/ directory within the
same.¹  Unlike ncov-ingest, however, forecasts-ncov publishes files
under its own name at <s3://nextstrain-data/files/workflows/forecasts-ncov/>
instead of ncov's name.  This is why it appears as its own "pathogen", a
conceptual confusion and one that means it gets unnecessary access to
write <s3://nextstrain-data/forecasts_ncov….json>.

We could choose to better address this some day, if the pattern of
data-generating repos downstream of pathogen repos proliferates, by
conceptually separating the two things in this Terraform configuration.

Related-to: <https://github.com/nextstrain/private/issues/110>
Alternative-to: <https://github.com/nextstrain/infra/pull/15>