Allow avian-flu repo to access seasonal-flu resources
Since the central ingest workflow for all flu data is running in
the seasonal-flu repo, allow avian-flu to access seasonal-flu resources
to fetch final metadata/sequences files under the seasonal-flu prefix.

Doing it this way instead of allowing seasonal-flu to have access to
avian-flu since it's easier to list separate input S3 URLs then it is to
upload to different prefixes.

Resolves https://github.com/nextstrain/infra/issues/57